refactor(carRacing) : Divide CarRacingView to InputView.kt and ResultView.kt
