+++ src/main/kotlin/carracing/InputView.kt
@@ -1,6 +1,6 @@
 package carracing
 
-object CarRacingView {
+object InputView {
 
     fun getCarNames(): String {
         println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
@@ -27,21 +27,5 @@ object CarRacingView {
             throw CarRacingException("자연수를 입력해주세요.")
         }
         return count
-    }
-
-    fun showCarMovement(cars: MutableList<Car>) {
-        cars.forEach {
-            print(it.name + " : " + "-".repeat(it.position) + "\n")
-        }
-    }
-
-    fun showWinners(winners: List<Car>) {
-        val result = StringBuilder()
-        winners.forEach {
-            result.append(it.name).append(", ")
-        }
-        result.delete(result.length - 2, result.length)
-
-        print(result.toString() + "가 최종 우승했습니다.")
     }
 }
